third_party/WebKit/LayoutTests/fast/dom/HTMLLinkElement/link-preload-unused-expected.txt

CONSOLE WARNING: The resource nick.jpg?name=value was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it wasn't preloaded for nothing.
 This test verifies that unused preload resources (and only unused ones) issue a warning.
